Multi version support

+++ elastestbrowsers/build_containers.sh
@@ -1,8 +1,11 @@
 #!/bin/bash -x
 set -e
 
+# Load old releases versions
+. browsers_oldreleases
 
-FIREFOX_OLD_VERSIONS="56.0 57.0 58.0"
+# Use MODE=FULL to generate all containers versions
+MODE=${MODE:-NIGHTLY}
 
 WORKDIR=$PWD/workdir
 [ -d $WORKDIR ] || mkdir -p $WORKDIR
@@ -24,7 +27,7 @@ cp -p workdir/selenoid_linux_amd64 firefox/image/selenoid/selenoid_linux_amd64
 # Firefox
 pushd firefox
 sed "s/VERSION/$FIREFOX_VER/g" image/selenoid/browsers.json.templ > image/selenoid/browsers.json
-docker build --build-arg VERSION=$FIREFOX_PKG -t elastestbrowsers/firefox:$FIREFOX_VER -t elastestbrowsers/firefox:latest -f Dockerfile.firefox .
+docker build --build-arg VERSION=$FIREFOX_PKG  -t elastestbrowsers/firefox:$FIREFOX_VER -t elastestbrowsers/firefox:latest -f Dockerfile.firefox .
 rm image/selenoid/browsers.json
 
 # Firefox Beta
@@ -36,15 +39,17 @@ rm image/selenoid/browsers.json.beta
 sed "s/VERSION/nightly/g" image/selenoid/browsers.json.templ > image/selenoid/browsers.json.nightly
 docker build -t elastestbrowsers/firefox:nightly -f Dockerfile.firefox.nightly .
 rm image/selenoid/browsers.json.nightly
-popd
 
 # Firefox old versions
-for V in $FIREFOX_OLD_VERSIONS
-do
-sed "s/VERSION/$V/g" image/selenoid/browsers.json.templ > image/selenoid/browsers.json
-docker build --build-arg VERSION=$V -t elastestbrowsers/firefox:$V -f Dockerfile.firefox .
-rm image/selenoid/browsers.json
-done
+if [ "$MODE" == "FULL" ]; then
+	for V in $FIREFOX_OLD_VERSIONS
+	do
+		sed "s/VERSION/$V/g" image/selenoid/browsers.json.templ > image/selenoid/browsers.json
+		docker build --build-arg VERSION=$V -t elastestbrowsers/firefox:$V -f Dockerfile.firefox.older_releases .
+		rm image/selenoid/browsers.json
+	done
+fi
+popd
 
 # cleaning
 rm firefox/image/selenoid/geckodriver
@@ -69,6 +74,36 @@ rm image/selenoid/browsers.json.beta
 sed "s/VERSION/unstable/g" image/selenoid/browsers.json.templ > image/selenoid/browsers.json.unstable
 docker build -t elastestbrowsers/chrome:unstable -f Dockerfile.chrome.unstable .
 rm image/selenoid/browsers.json.unstable
+
+# Chrome old versions
+if [ "$MODE" == "FULL" ]; then
+	for V in $CHROME_OLD_VERSIONS
+	do
+		rm -Rfv image/selenoid/chromedriver
+		sed "s/VERSION/$V/g" image/selenoid/browsers.json.templ > image/selenoid/browsers.json
+		TAG_VER=$(echo $V | cut -d"." -f1,2)
+		case $TAG_VER in
+			"60.0")
+				CHROMEDRIVER=2.33
+				;;
+			"61.0" | "62.0")
+				CHROMEDRIVER=2.34
+				;;
+			"63.0")
+				CHROMEDRIVER=2.36
+				;;
+			*)
+				CHROMEDRIVER=2.37
+				;;
+		esac
+		wget -O $WORKDIR/chromedriver.zip https://chromedriver.storage.googleapis.com/$CHROMEDRIVER/chromedriver_linux64.zip && \
+		unzip $WORKDIR/chromedriver.zip -d image/selenoid && \
+		rm -Rfv /chromedriver.zip
+		docker build --build-arg VERSION=$V -t elastestbrowsers/chrome:$TAG_VER -f Dockerfile.chrome.older_releases .
+		rm image/selenoid/browsers.json
+	done
+fi 
+
 popd
 
 # cleaning
